MBS-11502 extract CrossDevice models to test-summary
