+++ subprojects/gradle/test-summary/src/main/kotlin/com/avito/test/summary/TestSummarySender.kt
@@ -3,8 +3,8 @@ package com.avito.test.summary
 import com.avito.logger.LoggerFactory
 import com.avito.logger.create
 import com.avito.reportviewer.ReportsApi
-import com.avito.reportviewer.model.CrossDeviceSuite
 import com.avito.reportviewer.model.ReportCoordinates
+import com.avito.reportviewer.model.SimpleRunTest
 import com.avito.reportviewer.model.Team
 import com.avito.slack.ConjunctionMessagePredicate
 import com.avito.slack.CoroutinesSlackBulkSender
@@ -17,8 +17,13 @@ import com.avito.slack.SlackMessageUpdaterWithThreadMark
 import com.avito.slack.TextContainsStringCondition
 import com.avito.slack.model.SlackChannel
 import com.avito.slack.model.SlackSendMessageRequest
+import com.avito.test.model.TestStatus
 import com.avito.test.summary.compose.SlackSummaryComposer
 import com.avito.test.summary.compose.SlackSummaryComposerImpl
+import com.avito.test.summary.model.CrossDeviceRunTest
+import com.avito.test.summary.model.CrossDeviceStatus
+import com.avito.test.summary.model.CrossDeviceSuite
+import com.avito.test.summary.model.FailureOnDevice
 
 internal interface TestSummarySender {
 
@@ -65,10 +70,12 @@ internal class TestSummarySenderImpl(
     private val slackEmojiProvider = SlackEmojiProvider()
 
     override fun send() {
-        reportsApi.getCrossDeviceTestData(reportCoordinates).fold(
-            { suite -> send(suite, requireNotNull(reportsApi.tryGetId(reportCoordinates))) },
-            { throwable -> logger.critical("Can't get suite report", throwable) }
-        )
+        reportsApi.getTestsForRunId(reportCoordinates)
+            .map { toCrossDeviceTestData(it) }
+            .fold(
+                { suite -> send(suite, requireNotNull(reportsApi.tryGetId(reportCoordinates))) },
+                { throwable -> logger.critical("Can't get suite report", throwable) }
+            )
     }
 
     private fun send(suite: CrossDeviceSuite, reportId: String) {
@@ -134,5 +141,50 @@ internal class TestSummarySenderImpl(
                 null
             }
         )
+    }
+
+    private fun toCrossDeviceTestData(testData: List<SimpleRunTest>): CrossDeviceSuite {
+        return testData
+            .groupBy { it.name }
+            .map { (testName, runs) ->
+                val status: CrossDeviceStatus = when {
+                    runs.any { it.status is TestStatus.Lost } ->
+                        CrossDeviceStatus.LostOnSomeDevices
+
+                    runs.all { it.status is TestStatus.Skipped } ->
+                        CrossDeviceStatus.SkippedOnAllDevices
+
+                    runs.all { it.status is TestStatus.Manual } ->
+                        CrossDeviceStatus.Manual
+
+                    /**
+                     * Успешным прогоном является при соблюдении 2 условий:
+                     *  - Все тесты прошли (имеют Success статус)
+                     *  - Есть пропущенные тесты (скипнули на каком-то SDK например),
+                     *    но все остальные являются успешными (как минимум 1)
+                     */
+                    runs.any { it.status is TestStatus.Success } &&
+                        runs.all { it.status is TestStatus.Success || it.status is TestStatus.Skipped } ->
+                        CrossDeviceStatus.Success
+
+                    runs.all { it.status is TestStatus.Failure } ->
+                        CrossDeviceStatus.FailedOnAllDevices(runs.deviceFailures())
+
+                    runs.any { it.status is TestStatus.Failure } ->
+                        CrossDeviceStatus.FailedOnSomeDevices(runs.deviceFailures())
+
+                    else ->
+                        CrossDeviceStatus.Inconsistent
+                }
+                CrossDeviceRunTest(testName, status)
+            }
+            .let { CrossDeviceSuite(it) }
+    }
+
+    private fun List<SimpleRunTest>.deviceFailures(): List<FailureOnDevice> {
+        return this.filter { it.status is TestStatus.Failure }
+            .map {
+                FailureOnDevice(it.deviceName, (it.status as TestStatus.Failure).verdict)
+            }
     }
 }
