+++ subprojects/gradle/test-summary/src/main/kotlin/com/avito/test/summary/analysis/CrossDeviceSuteExtensions.kt
@@ -1,8 +1,8 @@
 package com.avito.test.summary.analysis
 
-import com.avito.reportviewer.model.CrossDeviceSuite
-import com.avito.reportviewer.model.FailureOnDevice
-import com.avito.reportviewer.model.HasFailures
+import com.avito.test.summary.model.CrossDeviceSuite
+import com.avito.test.summary.model.FailureOnDevice
+import com.avito.test.summary.model.HasFailures
 
 internal fun CrossDeviceSuite.analyzeFailures(): Map<String, List<FailureOnDevice>> {
     return this.crossDeviceRuns
