+++ subprojects/gradle/test-summary/src/main/kotlin/com/avito/test/summary/compose/SlackSummaryComposer.kt
@@ -1,9 +1,9 @@
 package com.avito.test.summary.compose
 
 import com.avito.android.Result
-import com.avito.reportviewer.model.CrossDeviceSuite
 import com.avito.reportviewer.model.ReportCoordinates
 import com.avito.reportviewer.model.Team
+import com.avito.test.summary.model.CrossDeviceSuite
 
 internal interface SlackSummaryComposer {
