+++ subprojects/gradle/test-summary/src/main/kotlin/com/avito/test/summary/model/CrossDeviceRunTest.kt
@@ -1,4 +1,4 @@
-package com.avito.reportviewer.model
+package com.avito.test.summary.model
 
 import com.avito.test.model.TestName
