+++ subprojects/gradle/test-summary/src/main/kotlin/com/avito/test/summary/model/CrossDeviceStatus.kt
@@ -1,4 +1,4 @@
-package com.avito.reportviewer.model
+package com.avito.test.summary.model
 
 public sealed class CrossDeviceStatus {
     public object Success : CrossDeviceStatus()
