+++ subprojects/gradle/test-summary/src/main/kotlin/com/avito/test/summary/model/CrossDeviceSuite.kt
@@ -1,7 +1,9 @@
-package com.avito.reportviewer.model
+package com.avito.test.summary.model
 
 import com.avito.math.Percent
 import com.avito.math.percentOf
+import com.avito.reportviewer.model.Team
+import com.avito.reportviewer.model.team
 
 public data class CrossDeviceSuite(val crossDeviceRuns: List<CrossDeviceRunTest>) {
