+++ subprojects/gradle/test-summary/src/testFixtures/kotlin/com/avito/test/summary/model/StubCrossDeviceRunTest.kt
@@ -1,0 +1,14 @@
+package com.avito.test.summary.model
+
+import com.avito.test.model.TestName
+import com.avito.test.model.createStubInstance
+
+public fun CrossDeviceRunTest.Companion.createStubInstance(
+    name: TestName = TestName.createStubInstance(),
+    status: CrossDeviceStatus = CrossDeviceStatus.Success,
+): CrossDeviceRunTest {
+    return CrossDeviceRunTest(
+        name = name,
+        status = status
+    )
+}
