+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/ReportsFetchApi.kt
@@ -1,7 +1,6 @@
 package com.avito.reportviewer
 
 import com.avito.android.Result
-import com.avito.reportviewer.model.CrossDeviceSuite
 import com.avito.reportviewer.model.Report
 import com.avito.reportviewer.model.ReportCoordinates
 import com.avito.reportviewer.model.SimpleRunTest
@@ -19,10 +18,6 @@ public interface ReportsFetchApi {
     /**
      * RunTest.List
      * получение краткого списка результатов тестов по запуску
-     *
-     * todo remove SimpleRunTest from public api
      */
     public fun getTestsForRunId(reportCoordinates: ReportCoordinates): Result<List<SimpleRunTest>>
-
-    public fun getCrossDeviceTestData(reportCoordinates: ReportCoordinates): Result<CrossDeviceSuite>
 }
