--- subprojects/test-runner/report-viewer/src/testFixtures/kotlin/com/avito/reportviewer/model/StubCrossDeviceRunTest.kt
@@ -1,8 +1,0 @@
-package com.avito.reportviewer.model
-
-import com.avito.test.model.TestName
-
-public fun CrossDeviceRunTest.Companion.createStubInstance(
-    name: TestName = TestName("com.test.Test", "test"),
-    status: CrossDeviceStatus = CrossDeviceStatus.Success
-): CrossDeviceRunTest = CrossDeviceRunTest(name, status)
