+++ subprojects/test-runner/test-model/build.gradle.kts
@@ -1,4 +1,5 @@
 plugins {
     id("convention.kotlin-jvm")
     id("convention.publish-kotlin-library")
+    id("convention.test-fixtures")
 }
