+++ subprojects/test-runner/test-model/src/testFixtures/kotlin/com/avito/test/model/StubTestName.kt
@@ -1,0 +1,9 @@
+package com.avito.test.model
+
+public fun TestName.Companion.createStubInstance(
+    className: String = "com.avito.Test",
+    methodName: String = "test"
+): TestName = TestName(
+    className = className,
+    methodName = methodName
+)
